Add GlobalsContainer to store global state
